Refactor: Improve code structure and add support modules
This commit restructures the NAS-TAS clustering and regime analysis code. It introduces new support modules for configuration, memory management, and algorithms, enhancing modularity and maintainability. The label fusion service is completely refactored for clarity and efficiency. Import handling is also improved with centralized management and fallback mechanisms.
