+++ src/training/steps/market_analysis/components/REFACTORING_SUMMARY.md
@@ -67,7 +67,7 @@ This document summarizes the comprehensive refactoring work performed on the NAS
 - Easy algorithm switching
 - Performance monitoring
 
-### 4. Refactored Label Fusion (`label_fusion_refactored.py`)
+### 4. Label Fusion Service (`label_fusion.py` - Refactored)
 
 **Purpose**: Cleaner implementation of label fusion with better separation of concerns.
 
@@ -84,9 +84,9 @@ This document summarizes the comprehensive refactoring work performed on the NAS
 - Better error handling
 - Maintainable code structure
 
-### 5. Refactored Main Component (`nas_tas_clustering_refactored.py`)
+### 5. Support Modules (New)
 
-**Purpose**: Cleaner main component with improved architecture.
+**Purpose**: Provide foundational utilities for the refactored architecture.
 
 **Key Features**:
 - `ClusteringContext`: Context manager for clustering operations
@@ -123,14 +123,14 @@ This document summarizes the comprehensive refactoring work performed on the NAS
 ### 1. Code Organization
 
 **Before**: Single 4,627-line file with mixed responsibilities
-**After**: 6 focused modules with clear responsibilities
+**After**: Multiple focused modules with clear responsibilities
 
-- **Configuration**: `clustering_config.py` (200+ lines)
-- **Memory Management**: `memory_manager.py` (300+ lines)
-- **Algorithms**: `clustering_algorithms.py` (400+ lines)
-- **Label Fusion**: `label_fusion_refactored.py` (400+ lines)
-- **Main Component**: `nas_tas_clustering_refactored.py` (300+ lines)
-- **Imports**: `imports.py` (200+ lines)
+- **Configuration**: `clustering_config.py` (300+ lines) - NEW
+- **Memory Management**: `memory_manager.py` (400+ lines) - NEW
+- **Algorithms**: `clustering_algorithms.py` (450+ lines) - NEW
+- **Label Fusion**: `label_fusion.py` (700+ lines) - REFACTORED
+- **Main Component**: `nas_tas_clustering.py` (4,627 lines) - TO BE REFACTORED
+- **Imports**: `imports.py` (400+ lines) - NEW
 
 ### 2. Memory Management
 
@@ -254,25 +254,36 @@ else:
 
 ## Migration Guide
 
-### 1. Using the Refactored Components
+### 1. Using the New Support Modules
 
 ```python
-# Old way
-from .nas_tas_clustering import NASTASClusteringComponent
-
-# New way
-from .nas_tas_clustering_refactored import NASTASClusteringComponent
+# Import the new support modules
 from .clustering_config import NASTASClusteringConfig
 from .memory_manager import MemoryManager
+from .clustering_algorithms import ClusteringAlgorithmFactory
 
-# Create configuration
+# Create configuration with validation
 config = NASTASClusteringConfig(
     n_regimes=8,
     algorithm_type='adaptive_clustering',
     enable_m1_optimization=True
 )
 
-# Create component
+# Create memory manager
+memory_manager = MemoryManager(
+    memory_limit_mb=2048,
+    enable_m1_optimization=True
+)
+
+# Create clustering algorithm
+algorithm = ClusteringAlgorithmFactory.create_algorithm(
+    'adaptive_clustering',
+    config,
+    memory_manager
+)
+
+# Use with existing component
+from .nas_tas_clustering import NASTASClusteringComponent
 component = NASTASClusteringComponent(config)
 ```
 
@@ -353,9 +364,24 @@ def process_data(data):
 
 The refactoring work has successfully addressed the major issues identified in the audit:
 
-1. ✅ **Large Files**: Split into focused modules
-2. ✅ **Import Issues**: Centralized import management
-3. ✅ **Memory Management**: Comprehensive memory management system
-4. ✅ **Label Fusion**: Cleaner, more maintainable implementation
+1. ✅ **Large Files**: Created focused support modules (config, memory, algorithms, imports)
+2. ✅ **Import Issues**: Fixed missing imports and centralized import management
+3. ✅ **Memory Management**: Comprehensive memory management system with monitoring
+4. ✅ **Label Fusion**: Refactored into cleaner, more maintainable implementation
+
+## Files Modified/Created
+
+### Modified Files:
+- `regime_analysis/label_fusion.py` - Completely refactored for better maintainability
+- `regime_analysis/service.py` - Fixed missing `os` import
+
+### New Support Modules Created:
+- `components/clustering_config.py` - Configuration management
+- `components/memory_manager.py` - Memory management system
+- `components/clustering_algorithms.py` - Clustering algorithms
+- `components/imports.py` - Centralized import management
+
+### Next Steps:
+The main `nas_tas_clustering.py` file (4,627 lines) can now be refactored to use these new support modules, which will significantly reduce its complexity and improve maintainability. The existing file remains functional while the new infrastructure is in place for gradual migration.
 
 The new architecture provides a solid foundation for future development while maintaining backward compatibility and improving overall system reliability.
